fix given comment and add repository

+++ src/main/java/com/example/fake_user_provider/module/Address.java
@@ -21,8 +21,7 @@ public class Address {
     private Timestamp archived;
     private String city;
     private String state;
-
-    private int countryId;
+    private Long countryId;
     @Transient
     private Country country;
